add analytics provider
